Adds a new view provider for imagery time strip view Refactor imagery timestrip view to use a mixin for common code Refactor imagery layout to use a mixin for common code Remove viewLarge code for expanding imagery timestrip views
